# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API for phpMyFAQ 4.1",
    "description": "phpMyFAQ includes a REST API and offers APIs for various services like fetching the phpMyFAQ version or doing a search against the phpMyFAQ installation.",
    "contact": {
      "name": "phpMyFAQ Team",
      "email": "support@phpmyfaq.de"
    },
    "license": {
      "name": "Mozilla Public Licence 2.0",
      "url": "https://www.mozilla.org/MPL/2.0/"
    },
    "version": "3.1"
  },
  "servers": [
    {
      "url": "https://localhost",
      "description": "Local dockerized server"
    }
  ],
  "paths": {
...
    "/api/v3.1/open-questions": {
      "get": {
        "tags": ["Public Endpoints"],
        "operationId": "getOpenQuestions",
        "responses": {
          "200": {
            "description": "Returns the open questions for the given language provided by \"Accept-Language\".",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the open questions.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "\n        [\n            {\n                \"id\": 1,\n                \"lang\": \"en\",\n                \"username\": \"phpMyFAQ User\",\n                \"email\": \"user@example.org\",\n                \"categoryId\": 3,\n                \"question\": \"Foo? Bar? Baz?\",\n                \"created\": \"20190106180429\",\n                \"answerId\": 0,\n                \"isVisible\": \"N\"\n              }\n        ]"
              }
            }
          },
          "404": {
            "description": "If no open questions are stored.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the open questions.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": []
              }
            }
          }
        }
      }
    },
...
    "/api/v3.1/question": {
      "post": {
        "tags": ["Endpoints with Authentication"],
        "operationId": "createQuestion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["category-id", "question", "author", "email"],
                "properties": {
                  "category-id": {
                    "type": "integer"
                  },
                  "question": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "type": "object"
              },
              "example": "{\n                \"category-id\": \"1\",\n                \"question\": \"Is this the world we created?\",\n                \"author\": \"Freddie Mercury\",\n                \"email\": \"freddie.mercury@example.org\"\n            }"
            }
          }
        },
        "responses": {
          "201": {
            "description": "Used to add a new question in one existing category.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"stored\": true }"
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3.1/register": {
      "post": {
        "tags": ["Endpoints with Authentication"],
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["username", "fullname", "email", "is-visible"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "fullname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "is-visible": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              },
              "example": "{\n                \"username\": \"ada\",\n                \"fullname\": \"Ada Lovelace\",\n                \"email\": \"ada.lovelace@example.org\",\n                \"is-visible\": false\n            }"
            }
          }
        },
        "responses": {
          "201": {
            "description": "If \"username\", \"fullname\", \"email\", and \"is-visible\" combination is correct.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"registered\": true, \"success\": \"User created.\"}"
              }
            }
          },
          "400": {
            "description": "If \"username\", \"fullname\", \"email\", and \"is-visible\" combination is not correct.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"registered\": false, \"error\": \"Error message\"}"
              }
            }
          },
          "409": {
            "description": "If the domain of the email address is not allowed.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"registered\": false, \"error\": \"The domain is not allowed.\"}"
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the question.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"registered\": false }"
              }
            }
          }
        }
      }
    },
...
  },
...
}